Add files via upload
temCalcs.py gives the functions that can be imported. It is not complete or optimal yet but it works.
temCalcsDemo.ipynb is a Jupyter Notebook file that demonstrates a selection of temCalc methods and classes.
the txt files are detector rotation calibrations for our TEM which can be imported and stored in Detector objects. They are used in the notebook.
I hope to one day incorporate this into a nice clean and easy user interface.